move schemas, etc. to their own file
